Updated CHANGELOG. Added full support for editing dates. You can now lauch reports from the contract pools

+++ client/trunk/ApplicationNodes/src/org/inventory/navigation/applicationnodes/pools/actions/DeletePoolAction.java
@@ -17,6 +17,7 @@ package org.inventory.navigation.applicationnodes.pools.actions;
 
 import java.awt.event.ActionEvent;
 import javax.swing.AbstractAction;
+import javax.swing.JOptionPane;
 import org.inventory.communications.CommunicationsStub;
 import org.inventory.core.services.api.notifications.NotificationUtil;
 import org.inventory.navigation.applicationnodes.pools.PoolNode;
@@ -44,12 +45,17 @@ public class DeletePoolAction extends AbstractAction {
 
     @Override
     public void actionPerformed(ActionEvent ev) {
-        if (com.deletePool(node.getPool().getOid())){
-            node.getParentNode().getChildren().remove(new Node[]{node});
-            NotificationUtil.getInstance().showSimplePopup("Success", NotificationUtil.INFO_MESSAGE, 
-                    java.util.ResourceBundle.getBundle("org/inventory/navigation/applicationnodes/Bundle").getString("LBL_DELETION_TEXT_OK"));
+        
+        if (JOptionPane.showConfirmDialog(null, "Are you sure you want to delete this pool? All contained elements will be deleted as well", 
+                "Warning", JOptionPane.OK_CANCEL_OPTION) == JOptionPane.OK_OPTION) {
+        
+            if (com.deletePool(node.getPool().getOid())){
+                node.getParentNode().getChildren().remove(new Node[]{node});
+                NotificationUtil.getInstance().showSimplePopup("Success", NotificationUtil.INFO_MESSAGE, 
+                        java.util.ResourceBundle.getBundle("org/inventory/navigation/applicationnodes/Bundle").getString("LBL_DELETION_TEXT_OK"));
+            }
+            else
+                NotificationUtil.getInstance().showSimplePopup("Error", NotificationUtil.ERROR_MESSAGE, com.getError());
         }
-        else
-            NotificationUtil.getInstance().showSimplePopup("Error", NotificationUtil.ERROR_MESSAGE, com.getError());
     }
 }
